+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v.Next
 
+- feat: Automatic Releasing via CI
+- feat: Navigation Instrumentation for Expo Router and React Native Router.
+- feat: Create `HoneycombReactNativeSDK` and `HoneycombReactNativeOptions`.
 - feat: Uncaught Exception handler
 - feat: Create `HoneycombReactNativeSDK` and `HoneycombReactNativeOptions`.
 - chore: Need to specify that this is a public npm package.